refactor code

+++ src/file/file.service.ts
@@ -1,9 +1,14 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import FileEntity from './entity/file.entity';
 import { FileMapper } from './file.mapper';
 import File from './interface/file.interface';
+import { v2 as cloudinary } from 'cloudinary';
 
 @Injectable()
 export class FileService {
@@ -39,7 +44,22 @@ export class FileService {
   }
 
   async remove(id: string): Promise<boolean> {
-    const result = await this.fileRepository.delete(id);
-    return result.affected !== 0; // true if a row was deleted, false otherwise
+    const file = await this.fileRepository.findOne({ where: { id } });
+    if (!file) {
+      throw new NotFoundException(`File with id ${id} not found`);
+    }
+
+    try {
+      // Delete file in cloud
+      if (file.save_path) {
+        await cloudinary.uploader.destroy(file.save_path);
+      }
+
+      const result = await this.fileRepository.delete(id);
+      return result.affected !== 0;
+    } catch (error) {
+      console.error('Error removing file:', error);
+      throw new Error('Cannot remove file');
+    }
   }
 }
